% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessing_mutating.R
\name{AccessMutate.JuliaProxy}
\alias{AccessMutate.JuliaProxy}
\alias{$.JuliaStructProxy}
\alias{$<-.JuliaStructProxy}
\alias{[.JuliaProxy}
\alias{[<-.JuliaProxy}
\alias{[[.JuliaArrayProxy}
\alias{[[<-.JuliaArrayProxy}
\alias{[[.JuliaStructProxy}
\alias{[[<-.JuliaStructProxy}
\alias{length.JuliaArrayProxy}
\alias{dim.JuliaArrayProxy}
\title{Access or mutate Julia objects}
\usage{
\method{$}{JuliaStructProxy}(x, name)

\method{$}{JuliaStructProxy}(x, name) <- value

\method{[}{JuliaProxy}(x, ...)

\method{[}{JuliaProxy}(x, i, j, k) <- value

\method{[[}{JuliaArrayProxy}(x, ...)

\method{[[}{JuliaArrayProxy}(x, i, j, k) <- value

\method{[[}{JuliaStructProxy}(x, name)

\method{[[}{JuliaStructProxy}(x, name) <- value

\method{length}{JuliaArrayProxy}(x)

\method{dim}{JuliaArrayProxy}(x)
}
\arguments{
\item{x}{a Julia proxy object}

\item{name}{a character string (TODO symbol) giving the name of a property
of a struct type or a key in a Julia dictionary (type \code{AbstractDict})}

\item{value}{a suitable replacement value.
When replacing a range of elements in an array type, it is possible to
replace multiple elements with single elements. In all other cases,
the length of the replacement must match the length of elements to replace.}

\item{i, j, k, ...}{indexes for specifying the elements to extract or replace}
}
\description{
Extract or replace parts of Julia proxy objects
}
\examples{
#TODO
#juliaEval("bla")
}
