% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaConnectoR-package.R
\name{EnvVars-JuliaConnectoR}
\alias{EnvVars-JuliaConnectoR}
\alias{JULIA_BINDIR}
\alias{JULIACONNECTOR_JULIAENV}
\alias{JULIACONNECTOR_SERVER}
\title{Environment variables used by the \pkg{JuliaConnectoR}}
\description{
There are some environment variables which can be used to deviate from the
default behavior of the package.
To have an effect, these environment variables must be set before a Julia
connection is established, i.e., before the first call to Julia or before a
call to  \code{\link{startJuliaServer}}.
All the variables are optional.
}
\details{
The environment variables that are used in the package are listed below:
\describe{
\item{\env{JULIA_BINDIR}:}{If this variable is set to the path of the Julia \code{bin} directory
  before connecting to Julia, the corresponding Julia installation will be used.
  By using this variable, it is possible to use a different Julia version
  than the one in the system \env{PATH}.
  (You can find the correct path to the \code{bin} directory
  of Julia by evaluating the expression \code{Sys.BINDIR} within Julia.)}
\item{\env{JULIACONNECTOR_JULIAENV}:}{Specify environment variables only for
   Julia.
   (This does not work on Windows and the variable is ignored there.)
   This allows, e.g., to set
   the LD_LIBRARY_PATH variable to a different value for Julia than for R.
   The value can be any R code that defines variables, e.g.,
   \code{"LD_LIBRARY_PATH=''"} is a valid value.
   On Linux, Julia is started with an empty \code{LD_LIBRARY_PATH}
   by default as the \code{LD_LIBRARY_PATH} required by R may be incompatible
   with Julia. If the \code{LD_LIBRARY_PATH} needs to be set to a different
   value, this can be done via the \code{JULIACONNECTOR_JULIAENV} variable.}
\item{\env{JULIACONNECTOR_JULIAOPTS}:}{Set start-up options for Julia.
   As an example, consider specifying the project environment and enabling
   code coverage when starting Julia. This can be achieved by setting
   the environment variable to
   \code{"--project=/path/to/project --code-coverage"}.}
\item{\env{JULIACONNECTOR_SERVER}:}{Specifies the server address of a
  (running) Julia server that the R process can connect to.
  A possible example value is "localhost:11980", specifying host and port.
  The function \code{\link{startJuliaServer}} sets this variable and
  communicates the location of the server to child processes with it.
  Due to security concerns, the Julia server accepts only connections from
  the same machine and connecting to remote machines is currently not
  possible.}
}
}
