% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{juliaCall}
\alias{juliaCall}
\title{Calls a Julia function.}
\usage{
juliaCall(name, ...)
}
\arguments{
\item{name}{name of the Julia function}

\item{...}{parameters handed to the function. Will be translated
to Julia data structures}
}
\value{
the value returned from Julia,
translated to an R data structure
}
\description{
Calls a Julia function.
}
