% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{juliaEval}
\alias{juliaEval}
\title{Evaluate a Julia expression}
\usage{
juliaEval(expr)
}
\arguments{
\item{expr}{Julia expression as a one-element character vector}
}
\value{
The value returned from Julia, translated to an R data structure.
If Julia returns \code{nothing}, an invisible \code{NULL} is returned.
}
\description{
This function evaluates a Julia expression in Julia
and translates the result back to R.
}
\details{
If the expression needs arguments, consider using \code{juliaLet}
instead.
}
\examples{
juliaEval("1 + 2")
juliaEval('using Pkg; Pkg.add("BoltzmannMachines")')
juliaEval('using Random; Random.seed!(5);')
\dontshow{
JuliaConnectoR:::stopJulia()
}
}
