% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{juliaFun}
\alias{juliaFun}
\title{Wrap a Julia function in an R function}
\usage{
juliaFun(name)
}
\arguments{
\item{name}{the name of the Julia function}
}
\description{
Creates an R function that will call the Julia function with the given name
when it is called. The returned function can also be passed as function argument
to Julia functions.
}
\examples{
juliaSqrt <- juliaFun("sqrt")
juliaSqrt(2)
juliaCall("map", juliaSqrt, c(1,4,9))
}
