% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{juliaGet}
\alias{juliaGet}
\title{Translate a Julia proxy object to an R object}
\usage{
juliaGet(x)
}
\arguments{
\item{x}{a reference to a Julia object}
}
\description{
R objects of class \code{JuliaProxy} are references to Julia objects in the Julia session.
These R objects are also called "proxy objects".
With this function it is possible to translate these objects into R objects.
}
\details{
If the corresponding Julia objects do not contain external references,
translated objects can also saved in R and safely be restored in Julia.

Modifying objects is possible and changes in R will be translated back to Julia.
}
