% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{juliaSetupOk}
\alias{juliaSetupOk}
\title{Check Julia setup}
\usage{
juliaSetupOk()
}
\value{
\code{TRUE} if the Julia setup is OK; otherwise \code{FALSE}
}
\description{
Checks that Julia can be started and that the Julia version is at least 1.0.
}
