% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaConnectoR-package.R
\name{Setup-JuliaConnectoR}
\alias{Setup-JuliaConnectoR}
\title{Julia setup}
\description{
Julia must be installed separately from the \pkg{JuliaConnectoR} package.
You can download and install Julia from \href{https://julialang.org/downloads/}{https://julialang.org/downloads/}.
After the installation, ensure that the Julia executable is in the system search \env{PATH} or that the
\env{JULIA_BINDIR} environment variable is set to the \code{bin} directory of
the Julia installation.
If the \env{JULIA_BINDIR} variable is set, it takes precedence over
looking in the executable path.
By setting the \env{JULIA_BINDIR} variable before starting Julia,
it is therefore possible to use different installations of Julia on the
same machine without having to change the executable path.
}
