% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{stopJulia}
\alias{stopJulia}
\title{Stop the connection to Julia}
\usage{
stopJulia()
}
\description{
This ends the connection to Julia. Julia will stop if no R process is
connected and the session is then terminated.
}
