% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overloading.R
\name{AccessOrMutate.JuliaObject}
\alias{AccessOrMutate.JuliaObject}
\alias{$.JuliaStruct}
\alias{$<-.JuliaStruct}
\alias{[.JuliaObject}
\alias{[<-.JuliaObject}
\alias{[[.JuliaArray}
\alias{[[<-.JuliaArray}
\alias{[[.JuliaStruct}
\alias{[[<-.JuliaStruct}
\title{Extract or Replace Parts of Julia Objects}
\usage{
\method{$}{JuliaStruct}(jlref, name)

\method{$}{JuliaStruct}(jlref, name) <- value

\method{[}{JuliaObject}(jlref, ...)

\method{[}{JuliaObject}(jlref, i, j, k) <- value

\method{[[}{JuliaArray}(jlref, ...)

\method{[[}{JuliaArray}(jlref, i, j, k) <- value

\method{[[}{JuliaStruct}(jlref, name)

\method{[[}{JuliaStruct}(jlref, name) <- value
}
\description{
Extract or replace parts of Julia proxy objects
}
\examples{
#TODO
juliaEval("bla")
}
