% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{juliaEval}
\alias{juliaEval}
\title{Evaluate a Julia expression}
\usage{
juliaEval(expr)
}
\arguments{
\item{expr}{Julia expression as a one-element character vector}
}
\value{
the value of the expression, translated to R
}
\description{
This function evaluates a Julia expression in Julia
and translates the result back to R.
}
\details{
If the expression needs arguments, consider using \code{juliaLet}
instead.
}
\examples{
juliaEval("1 + 2")
juliaEval('using Pkg; Pkg.add("BoltzmannMachines")')
juliaEval('using Random; Random.seed!(5);')

}
