% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{juliaImport}
\alias{juliaImport}
\title{Load and import a Julia package via \code{import} statement}
\usage{
juliaImport(modulePath, alias = NULL, importInternal = FALSE)
}
\arguments{
\item{modulePath}{name of the package/module that is to be used,
or a relative module path.
Specifying a Julia module path like \code{.MyModule}
allows importing a module which does not correspond to a package,
but has been loaded in the \code{Main} module, e. g. by
\code{juliaCall("include", "path/to/MyModule.jl")}.
Additionally, via a path such as \code{SomePkg.SubModule},
a submodule of a package can be imported.}

\item{alias}{alternative prefix for the package
(useful e.g. to avoid naming collisions or for brevity)
If an alias is not explicitly specified, the name of the
package/module is used.}

\item{importInternal}{\code{logical} value, default \code{FALSE}.
Specifies whether unexported functions shall be imported.}
}
\description{
The specified package/module is loaded via \code{import} in Julia
and its functions are attached to the R search path.
This way, all functions (including constructors) exported by the
package are available in R under their name
prefixed with the module name plus "\code{.}", like in Julia.
}
\examples{
# Importing a package and using one of its exported functions
juliaImport("UUIDs")
juliaCall("string", UUIDs.uuid4())

# Functions that are not exported can be imported
# by specifying the argument "importInternal":
juliaImport("Pkg", importInternal = TRUE)
Pkg.status()

# Importing a module without a package
testModule <- system.file("examples", "TestModule1.jl",
                          package = "JuliaConnectoR")
# take a look at the file
writeLines(readLines(testModule))
# load in Julia
juliaCall("include", testModule)
# import in R
juliaImport(".TestModule1")
TestModule1.test1()

# Importing a submodule
testModule <- system.file("examples", "TestModule1.jl",
                          package = "JuliaConnectoR")
juliaCall("include", testModule)
juliaImport(".TestModule1.SubModule1")
# call exported function of submodule via module path
SubModule1.test2()
juliaImport(".TestModule1.SubModule1", alias = "Sub1")
# call exported function of submodule via alias
Sub1.test2()

\dontshow{
JuliaConnectoR:::stopJulia()
}
}
