% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{juliaUsing}
\alias{juliaUsing}
\title{Load and import a Julia package via \code{using} statement}
\usage{
juliaUsing(modulePath, alias = NULL, importInternal = FALSE)
}
\arguments{
\item{modulePath}{name of the package/module that is to be used,
or a relative module path.
Specifying a Julia module path like \code{.MyModule}
allows using a module which does not correspond to a package,
but has been loaded in the \code{Main} module, e. g. by
\code{juliaCall("include", "path/to/MyModule.jl")}.
Additionally, via a path such as \code{SomePkg.SubModule},
a submodule of a package can be imported.}

\item{alias}{alternative prefix for the package
(useful e.g. to avoid naming collisions or for brevity)
If an alias is not explicitly specified, the name of the
package/module is used.}

\item{importInternal}{\code{logical} value, default \code{FALSE}.
Specifies whether unexported functions shall be imported.}
}
\description{
The specified package/module is loaded via \code{using} in Julia
and its functions are attached to the R search path.
This way, all functions (including constructors) exported by the
package are available in R under their name, and under the name
prefixed with the module name plus "\code{.}".
}
\examples{
\dontrun{
# Using a package and one of its exported functions
juliaUsing("UUIDs")
juliaCall("string", uuid4())

# Functions that are not exported can be imported
# by specifying the argument "importInternal":
juliaUsing("Pkg", importInternal = TRUE)
Pkg.status()

# Using a module without a package
testModule <- system.file("examples", "TestModule1.jl",
                          package = "JuliaConnectoR")
# take a look at the file
writeLines(readLines(testModule))
# load in Julia
juliaCall("include", testModule)
# import via "using" in R
juliaUsing(".TestModule1")
# call exported function
test1()
# execute exported function via module name
TestModule1.test1()

# Using a submodule
testModule <- system.file("examples", "TestModule1.jl",
                          package = "JuliaConnectoR")
juliaCall("include", testModule)
juliaUsing(".TestModule1.SubModule1")
# call exported function of submodule with
test2()
# ... or with
SubModule1.test2()
}
}
