% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{startJuliaServer}
\alias{startJuliaServer}
\title{Start a server that may serves multiple clients (R processes)}
\usage{
startJuliaServer(port = 11980)
}
\arguments{
\item{port}{a hint for the port that is used by the server.
If it is not available, a different port is used.
The final port is returned invisibly.}
}
\value{
the port number (invisibly)
}
\description{
TODO some more info
}
